fix: remade from fun to class
